# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OT & Salary Calculator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SalarySettings {
  id: number;
  userId: string;
  /** Monthly base salary in THB */
  baseSalary: number;
  /** OT rate multiplier (e.g. 1.5 or 3.0) */
  otRate: number;
  /** Standard working hours per day (e.g. 8) */
  hoursPerDay: number;
  /** Standard working days per month (e.g. 26) */
  workingDaysPerMonth: number;
  updatedAt: string;
}

export interface SalarySettingsInput {
  baseSalary: number;
  otRate: number;
  hoursPerDay: number;
  workingDaysPerMonth: number;
}

export interface OtEntry {
  id: number;
  userId: string;
  /** Date of OT (YYYY-MM-DD) */
  date: string;
  /** Number of OT hours */
  hours: number;
  /** Type: weekday, weekend, holiday */
  otType: string;
  /** @nullable */
  note?: string | null;
  /** Calculated OT pay for this entry */
  otPay: number;
  createdAt: string;
}

export interface OtEntryInput {
  date: string;
  hours: number;
  otType: string;
  note?: string;
}

export interface OtEntryUpdate {
  date?: string;
  hours?: number;
  otType?: string;
  /** @nullable */
  note?: string | null;
}

export interface MonthlySummary {
  month: string;
  baseSalary: number;
  totalOtHours: number;
  totalOtPay: number;
  totalSalary: number;
  weekdayOtHours: number;
  weekendOtHours: number;
  holidayOtHours: number;
  entriesCount: number;
}

export interface YearlySummary {
  year: string;
  totalOtHours: number;
  totalOtPay: number;
  totalBaseSalary: number;
  totalSalary: number;
  monthlyBreakdown: MonthlySummary[];
}

export type ListOtEntriesParams = {
/**
 * Filter by month (YYYY-MM)
 */
month?: string;
};

export type GetMonthlySummaryParams = {
/**
 * Month to summarize (YYYY-MM)
 */
month: string;
};

export type GetYearlySummaryParams = {
/**
 * Year (YYYY)
 */
year: string;
};