# lib/api-zod/src/generated/types/monthlySummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OT & Salary Calculator API
 * OpenAPI spec version: 0.1.0
 */

export interface MonthlySummary {
  month: string;
  baseSalary: number;
  totalOtHours: number;
  totalOtPay: number;
  totalSalary: number;
  weekdayOtHours: number;
  weekendOtHours: number;
  holidayOtHours: number;
  entriesCount: number;
}