# lib/api-zod/src/generated/types/otEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OT & Salary Calculator API
 * OpenAPI spec version: 0.1.0
 */

export interface OtEntry {
  id: number;
  userId: string;
  /** Date of OT (YYYY-MM-DD) */
  date: string;
  /** Number of OT hours */
  hours: number;
  /** Type: weekday, weekend, holiday */
  otType: string;
  /** @nullable */
  note?: string | null;
  /** Calculated OT pay for this entry */
  otPay: number;
  createdAt: string;
}