# lib/api-zod/src/generated/types/otEntryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OT & Salary Calculator API
 * OpenAPI spec version: 0.1.0
 */

export interface OtEntryInput {
  date: string;
  hours: number;
  otType: string;
  note?: string;
}