# lib/api-zod/src/generated/types/otEntryUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OT & Salary Calculator API
 * OpenAPI spec version: 0.1.0
 */

export interface OtEntryUpdate {
  date?: string;
  hours?: number;
  otType?: string;
  /** @nullable */
  note?: string | null;
}