# lib/api-zod/src/generated/types/salarySettings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OT & Salary Calculator API
 * OpenAPI spec version: 0.1.0
 */

export interface SalarySettings {
  id: number;
  userId: string;
  /** Monthly base salary in THB */
  baseSalary: number;
  /** OT rate multiplier (e.g. 1.5 or 3.0) */
  otRate: number;
  /** Standard working hours per day (e.g. 8) */
  hoursPerDay: number;
  /** Standard working days per month (e.g. 26) */
  workingDaysPerMonth: number;
  updatedAt: string;
}