# lib/api-zod/src/generated/types/salarySettingsInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OT & Salary Calculator API
 * OpenAPI spec version: 0.1.0
 */

export interface SalarySettingsInput {
  baseSalary: number;
  otRate: number;
  hoursPerDay: number;
  workingDaysPerMonth: number;
}