# lib/api-zod/src/generated/types/yearlySummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OT & Salary Calculator API
 * OpenAPI spec version: 0.1.0
 */
import type { MonthlySummary } from './monthlySummary';

export interface YearlySummary {
  year: string;
  totalOtHours: number;
  totalOtPay: number;
  totalBaseSalary: number;
  totalSalary: number;
  monthlyBreakdown: MonthlySummary[];
}